requirements.txt
asgiref==3.10.0
Django==5.2.5
django-cors-headers==4.9.0
django-filter==25.2
djangorestframework==3.16.1
djangorestframework_simplejwt==5.5.1
pillow==12.0.0
PyJWT==2.10.1
setuptools==80.9.0
sqlparse==0.5.3
wheel==0.45.1
